requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
Django==4.0.2
django-bootstrap-v5==1.0.11
django-crispy-forms==1.14.0
django-model-utils==4.2.0
django-quill-editor==0.1.40
django-taggit==2.1.0
django-widget-tweaks==1.4.12
fontawesomefree==6.0.0
Pillow==9.0.1
soupsieve==2.3.1
sqlparse==0.4.2
tzdata==2021.5
